smalltalk_package.bbclass: Fix gst not finding sysroot kernel directory
A recipe inheriting smalltalk_package was failing due to gst being
unable to find Builtins.st in the kernel directory.

Using strace on gst while running in the OE env showed that gst looks
for the kernel directory in /kernel and in the tmp/work/smalltalk
directory where it was built, but it doesn't look in the native sysroot
of the recipe being built, were the kernel/Builtins.st file can be
found.